HotFix: PG to PG is empty bug (#130)
There was an inconsistency with the types returned from postgres and
those being passed in.

PG.fetch returned `DataFrame`
but 
PG.save expected `TypedDataFrame`

Resulted in a bug where data.is_empty did not exist. 

This PR aligns the types naively, but this will all be improved in a
follow up #129 (which assigns proper/non-empty types) to Postgres
dataframes.